requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
pyyaml>=6.0
requests>=2.31
python-multipart>=0.0.9
databricks-sdk>=0.74.0
